refact: move orderId state from current-burger-slice to ui-slice
